refactor(console): decompose the servers page map/rows
Two related splits of the servers page:
- Extract the server/site row decoration (regionsByName + serverRows +
  siteRows + merged sorted list) from ServersPage into a useFleetRows
  composable; move the ResourceRow type from ServerListPanel.vue into
  lib/serverMap.ts so the panel and composable share one definition.
- Lift the ~190-line server/cluster/plus hover-card template out of ServerMap
  into a presentational MapHoverCard child, leaving ServerMap to own just the
  pan/zoom viewport.

Same props/emits/slots throughout, so behaviour is unchanged.
