+++ dashboard/src/components/servers/ServerMap.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import { Badge, Button } from 'frappe-ui'
 import {
 	type CSSProperties,
 	computed,
@@ -8,6 +7,7 @@ import {
 	ref,
 	watch,
 } from 'vue'
+import MapHoverCard from '@/components/servers/MapHoverCard.vue'
 import ProviderAvatar from '@/components/servers/ProviderAvatar.vue'
 import WorldDots from '@/components/servers/WorldDots.vue'
 import {
@@ -399,10 +399,6 @@ function cancelHide(): void {
 	window.clearTimeout(hideT)
 }
 
-function canOpenBench(server: NonNullable<MapPin['server']>): boolean {
-	return props.allowOpen && server.status === 'Running' && !!server.gateway_url
-}
-
 function hideCard(): void {
 	window.clearTimeout(showT)
 	window.clearTimeout(hideT)
@@ -641,169 +637,20 @@ function clickNode(n: MapNode): void {
 				@mouseleave="leaveNode"
 				@click.capture="cardLocked = true"
 			>
-				<!-- Single VM (server or site): real mirror fields only. The IP/Plan/Version
-             rows are server-only and simply don't render for a site (fields absent). -->
-				<template v-if="card.node.type === 'server'">
-					<div class="flex items-start gap-2">
-						<div class="min-w-0 flex-1">
-							<div class="flex items-center gap-2">
-								<span class="truncate text-base font-semibold text-ink-gray-9"
-									>{{ card.node.pin.name }}</span
-								>
-								<Badge
-									:theme="card.node.pin.visual.badgeTheme"
-									variant="subtle"
-									size="sm"
-									:label="card.node.pin.visual.label"
-								/>
-							</div>
-							<div
-								v-if="card.node.pin.specs"
-								class="mt-0.5 truncate text-sm text-ink-gray-5"
-							>
-								{{ card.node.pin.specs }}
-							</div>
-						</div>
-						<div class="-mr-1.5 -mt-1" @click.stop>
-							<slot name="card-actions" :pin="card.node.pin" />
-						</div>
-					</div>
-					<div class="mt-3 flex items-baseline justify-between gap-3 text-sm">
-						<span class="shrink-0 font-medium text-ink-gray-8">Region</span>
-						<span class="truncate text-ink-gray-9"
-							>{{ card.node.pin.flag }} {{ card.node.pin.regionLabel }}</span
-						>
-					</div>
-					<div
-						v-if="card.node.pin.publicIpv4"
-						class="mt-2 flex items-baseline justify-between gap-3 text-sm"
-					>
-						<span class="shrink-0 font-medium text-ink-gray-8">IP</span>
-						<span class="truncate font-mono text-[13px] text-ink-gray-9"
-							>{{ card.node.pin.publicIpv4 }}</span
-						>
-					</div>
-					<div
-						v-if="card.node.pin.plan"
-						class="mt-2 flex items-baseline justify-between gap-3 text-sm"
-					>
-						<span class="shrink-0 font-medium text-ink-gray-8">Plan</span>
-						<span class="truncate text-ink-gray-9"
-							>{{ card.node.pin.plan }}</span
-						>
-					</div>
-					<div
-						v-if="card.node.pin.frappeVersion"
-						class="mt-2 flex items-baseline justify-between gap-3 text-sm"
-					>
-						<span class="shrink-0 font-medium text-ink-gray-8">Version</span>
-						<span class="truncate text-ink-gray-9"
-							>{{ card.node.pin.frappeVersion }}</span
-						>
-					</div>
-				</template>
-
-				<!-- Cluster: the servers at this spot -->
-				<template v-else-if="card.node.type === 'cluster'">
-					<div
-						class="flex items-center justify-between gap-2 px-1.5 pb-1 pt-0.5"
-					>
-						<span class="min-w-0 truncate text-xs font-medium text-ink-gray-5">
-							{{ card.node.members[0].flag }} {{ card.node.title }} ·
-							{{ card.node.members.length }}
-							servers
-						</span>
-						<button
-							v-if="allowCreate"
-							class="grid size-5 shrink-0 place-items-center rounded-md text-ink-gray-6 transition-colors hover:bg-surface-gray-2 hover:text-ink-gray-8 active:scale-95"
-							:title="`New server in ${card.node.title}`"
-							:aria-label="`New server in ${card.node.title}`"
-							@click="emit('new-server', card.node.members[0].cluster)"
-						>
-							<span class="lucide-plus size-3.5" />
-						</button>
-					</div>
-					<div
-						v-for="m in card.node.members"
-						:key="m.id"
-						class="group flex w-full items-center gap-2.5 rounded-lg p-1.5 transition-colors hover:bg-surface-gray-2"
-					>
-						<button
-							class="flex min-w-0 flex-1 items-center gap-2.5 text-left"
-							@click="emit('open', m.id)"
-						>
-							<span class="relative shrink-0">
-								<ProviderAvatar :provider="m.provider" :size="28" />
-								<span
-									class="absolute -bottom-px -right-px size-2.5 rounded-full border-2 border-[var(--surface-elevation-1)]"
-									:style="{ background: m.visual.dot }"
-								/>
-							</span>
-							<span class="min-w-0 flex-1">
-								<span class="block truncate text-sm font-medium text-ink-gray-8"
-									>{{ m.name }}</span
-								>
-								<span class="block truncate text-xs text-ink-gray-5"
-									>{{ m.specs }}</span
-								>
-							</span>
-						</button>
-						<!-- Server: open bench. Site: open its live URL. Same slot, per kind. -->
-						<button
-							v-if="m.kind === 'server' && m.server"
-							class="grid size-7 shrink-0 place-items-center rounded text-ink-gray-5 transition-opacity disabled:cursor-default disabled:opacity-30 enabled:opacity-0 enabled:hover:text-ink-gray-8 group-hover:enabled:opacity-100"
-							:disabled="!canOpenBench(m.server)"
-							title="Open bench"
-							aria-label="Open bench"
-							@click.stop="emit('open-server', m.server)"
-						>
-							<span class="lucide-arrow-up-right size-3.5" />
-						</button>
-						<button
-							v-else-if="m.site"
-							class="grid size-7 shrink-0 place-items-center rounded text-ink-gray-5 transition-opacity disabled:cursor-default disabled:opacity-30 enabled:opacity-0 enabled:hover:text-ink-gray-8 group-hover:enabled:opacity-100"
-							:class="{ '!opacity-100': openingSite === m.site.name }"
-							:disabled="!allowOpen || !m.site.url || openingSite === m.site.name"
-							title="Open site"
-							aria-label="Open site"
-							@click.stop="m.site.url && emit('open-site', m.site.name)"
-						>
-							<span
-								:class="openingSite === m.site.name ? 'lucide-loader-circle size-3.5 animate-spin' : 'lucide-arrow-up-right size-3.5'"
-							/>
-						</button>
-					</div>
-				</template>
-
-				<!-- Empty region: a direct path to create (markers never open cards) -->
-				<template v-else-if="card.node.type === 'plus'">
-					<div class="text-base font-semibold text-ink-gray-9">
-						No servers in this region
-					</div>
-					<div class="mt-0.5 text-sm text-ink-gray-5">
-						{{ card.node.title }}
-					</div>
-					<div class="mt-3 flex items-center gap-2">
-						<span class="text-sm text-ink-gray-6">Providers available</span>
-						<button
-							v-for="t in card.node.targets"
-							:key="t.id"
-							class="block shrink-0 rounded-full transition-transform duration-150 ease-out hover:scale-110 active:scale-95"
-							:title="`New server in ${t.flag} ${t.regionLabel}`"
-							@click="emit('new-server', t.id)"
-						>
-							<ProviderAvatar :provider="t.provider" :size="20" />
-						</button>
-					</div>
-					<Button
-						class="mt-3"
-						variant="subtle"
-						size="sm"
-						label="New server"
-						icon-left="lucide-plus"
-						@click="emit('new-server', card.node.targets[0].id)"
-					/>
-				</template>
+				<MapHoverCard
+					:node="card.node"
+					:allow-create="allowCreate"
+					:allow-open="allowOpen"
+					:opening-site="openingSite"
+					@open="emit('open', $event)"
+					@open-server="emit('open-server', $event)"
+					@open-site="emit('open-site', $event)"
+					@new-server="emit('new-server', $event)"
+				>
+					<template #card-actions="{ pin }">
+						<slot name="card-actions" :pin="pin" />
+					</template>
+				</MapHoverCard>
 			</div>
 		</Transition>
